Test get request on a certain post
send a get request on the post endpoint /api/post/1 to retrieve this post
which belongs to the user michael and assert if the respond contains the
char sequence "michael". If so, test is successful.

Note! Main Application must be running when running the test as its not
being initialised upon test run.